fatal: move enum definition to dedicated header
This allows enum values to be referenced in architecture `arch.h` files
without circular dependencies.
